requirements.txt
Flask
flask_cors
gunicorn
google-generativeai
requests
bs4